Include wrapped neighbours when counting live neighbours
To satisfy the rule 'A Cell who "comes to life" outside the board should
wrap at the other side of the board', cells at the edge of the board
need to include the cells on the opposite side of the board as their
neighbours.

Also let's ensure cells that are out of bounds are considered to have
zero live neighbours.